updated/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
billiard==4.2.2
cachetools==6.2.2
celery==5.5.3
certifi==2025.10.5
channels==4.3.1
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.7
django-auditlog==3.3.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
fastapi==0.118.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-core==2.5.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gunicorn==23.0.0
h11==0.16.0
httplib2==0.31.0
httptools==0.6.4
idna==3.10
kombu==5.5.4
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.10
pydantic_core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.14
websockets==15.0.1
whitenoise==6.11.0
